feat(views): adicionando validacoes de views e refatorando BancoView
