+++ src/br/com/ada/contas/ContaInvestimento.java
@@ -2,7 +2,6 @@ package br.com.ada.contas;
 
 import br.com.ada.clientes.Cliente;
 import br.com.ada.clientes.TipoCliente;
-import br.com.ada.validacoes.Verificar;
 
 public class ContaInvestimento extends Conta {
     public ContaInvestimento(Cliente cliente) {
