+++ src/br/com/ada/utilidades/InformacoesDeInput.java
@@ -1,0 +1,6 @@
+package br.com.ada.utilidades;
+
+public class InformacoesDeInput {
+    public static final int TAMANHO_MAXIMO_SENHA = 6;
+    public static final int TAMANHO_MINIMO_SENHA = 4;
+}
