+++ src/br/com/ada/utilidades/InformacoesDeTaxas.java
@@ -1,0 +1,6 @@
+package br.com.ada.utilidades;
+
+public class InformacoesDeTaxas {
+    public static final double TAXA_SAQUE_CLIENTE_FISICO = 0;
+    public static final double TAXA_SAQUE_CLIENTE_JURIDICO = 0;
+}
