+++ src/br/com/ada/validacoes/ValidarInput.java
@@ -1,0 +1,14 @@
+package br.com.ada.validacoes;
+
+public class ValidarInput {
+    public ValidacaoString validarString = new ValidacaoString();
+    public ValidarCnpj validarCnpj = new ValidarCnpj();
+    public ValidarCpf validarCpf = new ValidarCpf();
+
+    public ValidarNome validarNome = new ValidarNome();
+
+    public ValidarSenha validarSenha = new ValidarSenha();
+
+    public ValidarInput() {
+    }
+}
