+++ src/br/com/ada/validacoes/ValidarNome.java
@@ -1,0 +1,4 @@
+package br.com.ada.validacoes;
+
+public class ValidarNome extends ValidacaoString{
+}
