+++ src/br/com/ada/validacoes/ValidarSenha.java
@@ -1,0 +1,4 @@
+package br.com.ada.validacoes;
+
+public class ValidarSenha extends ValidacaoString{
+}
