+++ src/br/com/ada/views/View.java
@@ -1,15 +1,19 @@
 package br.com.ada.views;
 
+import br.com.ada.validacoes.ValidacaoString;
+import br.com.ada.validacoes.ValidarInput;
+
 import java.util.Scanner;
 
 public abstract class View {
     public Scanner sc = new Scanner(System.in);
+    public ValidarInput validarInput = new ValidarInput();
 
     public double getDouble () {
         double valor = 0;
 
         try {
-            valor = Double.parseDouble(sc.nextLine());
+            valor = Double.parseDouble(getString());
         } catch (Exception e) {
             System.out.println("Valor inválido, tente novamente.");
             getDouble();
@@ -21,7 +25,7 @@ public abstract class View {
         int valor = 0;
 
         try {
-            valor = Integer.parseInt(sc.nextLine());
+            valor = Integer.parseInt(getString());
         } catch (Exception e) {
             System.out.println("Valor inválido, tente novamente.");
             getInt();
@@ -31,6 +35,24 @@ public abstract class View {
     }
 
     public String getString () {
-        return sc.nextLine();
+        String string = sc.nextLine();
+        if(validarInput.validarString.validarVazio(string)){
+            return getString();
+        };
+        return string;
+    }
+
+    public String pedirCpf() {
+        System.out.println("Digite o seu cpf:");
+        String cpf =  getString();
+        return cpf;
+        //todo - validar cpf
+    }
+
+    public String pedirCnpj() {
+        System.out.println("Digite o seu cnpj:");
+        String cnpj =  getString();
+        return cnpj;
+        //todo - validar cnpj
     }
 }
