+++ src/br/com/ada/views/banco/MenuInciarView.java
@@ -1,0 +1,39 @@
+package br.com.ada.views.banco;
+
+import br.com.ada.views.View;
+
+public class MenuInciarView extends View {
+    private final MenuCadastroView menuCadastroView;
+    private final MenuLoginView menuLoginView;
+
+    public MenuInciarView(MenuCadastroView menuCadastroView, MenuLoginView menuLoginView) {
+        this.menuCadastroView = menuCadastroView;
+        this.menuLoginView = menuLoginView;
+    }
+
+    public void menuInicial() {
+        boolean sair = false;
+        do {
+            System.out.println("");
+            System.out.println("Bem vindo ao Banco");
+            System.out.println("Digite 1 para abrir conta.");
+            System.out.println("Digite 2 para acessar sua conta.");
+            System.out.println("Digite 3 para encerrar a sessão.");
+            int opcaoMenu = getInt();
+            switch (opcaoMenu) {
+                case 1:
+                    menuCadastroView.menuCadastro();
+                    break;
+                case 2:
+                    menuLoginView.menuLogin();
+                    break;
+                case 3:
+                    sair = true;
+                    break;
+                default:
+                    System.out.println("Opção inválida, por favor digite novamente");
+            }
+        } while(!sair);
+    }
+
+}
